requirements.txt
pyTelegramBotAPI
flask
mysql-connector-python
